Dev merge in main version 0.0.2

+++ app/Http/Controllers/RoleController.php
@@ -72,7 +72,7 @@ class RoleController extends Controller
 
     /**
      * delete one role
-     * * 200 [message]
+     * * 200 [message, delete]
      * * 401 [message]
      *
      * @param Request $request
@@ -81,7 +81,16 @@ class RoleController extends Controller
     public function delete(Request $request): JsonResponse
     {
         $role = Role::find($request->id);
+        if ($role->users()) {
+            return response()->json([
+                'message' => 'Des utilisateurs ont ce role, impossible de le supprimer',
+                'delete' => false,
+            ], 200);
+        }
         $role->delete();
-        return response()->json(['message' => 'Le role a été supprimé'], 200);
+        return response()->json([
+            'message' => 'Le role a été supprimé',
+            'delete' => true,
+        ], 200);
     }
 }
